% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phewas_category_overlap}
\alias{plot_phewas_category_overlap}
\title{Plot FDR signal overlap by phenotype group}
\usage{
plot_phewas_category_overlap(comparison)
}
\arguments{
\item{comparison}{A \code{phewas_comparison} from \code{\link[=compare_phewas]{compare_phewas()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot FDR signal overlap by phenotype group
}
