% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_phewas_bundle}
\alias{read_phewas_bundle}
\title{Read PheWAS results}
\usage{
read_phewas_bundle(directory)
}
\arguments{
\item{directory}{Result directory created by \code{\link[=write_phewas_bundle]{write_phewas_bundle()}}.}
}
\value{
A \code{phewas_result} object.
}
\description{
Read PheWAS results
}
